5th

+++ src/Data/Apartado.java
@@ -200,7 +200,7 @@ public class Apartado {
                         detalleVenta.clear();
                         while (detalle.next()) {
                             Articulo a = new Articulo();
-                            a.setCantidadVenta(detalle.getInt("cantidad"));
+                            a.setCantidadVenta(detalle.getFloat("cantidad"));
                             a.setCodigo(detalle.getString("codigo"));
                             a.setIdArticulo(detalle.getInt("id_articulo"));
                             a.setNombre(detalle.getString("nombre"));
@@ -288,7 +288,7 @@ public class Apartado {
                         detalleVenta.clear();
                         while (detalle.next()) {
                             Articulo a = new Articulo();
-                            a.setCantidadVenta(detalle.getInt("cantidad"));
+                            a.setCantidadVenta(detalle.getFloat("cantidad"));
                             a.setCodigo(detalle.getString("codigo"));
                             a.setIdArticulo(detalle.getInt("id_articulo"));
                             a.setNombre(detalle.getString("nombre"));
